+++ js-packages/@prestojs/ui-antd/src/widgets/IntegerWidget.tsx
@@ -17,7 +17,7 @@ type IntegerWidgetProps = WidgetProps<number, HTMLInputElement>;
  */
 function IntegerWidget(
     props: IntegerWidgetProps,
-    ref: React.RefObject<typeof InputNumber>
+    ref: React.RefObject<HTMLInputElement>
 ): React.ReactElement {
     const { input, ...rest } = props;
     return <InputNumber ref={ref} {...rest} {...input} />;